feat: split the webeval into two windows, a configuration under options, and a dockable window under windows tabs
