+++ src/gui/CMakeLists.txt
@@ -46,6 +46,7 @@ set(gui_SOURCES
         windows/registers/registersdock.cpp
         dialogs/savechanged/savechangeddialog.cpp
         dialogs/webeval/webevalconfigdialog.cpp
+        windows/webeval/webevaldock.cpp
         windows/editor/srceditor.cpp
         statictable.cpp
         windows/terminal/terminaldock.cpp
@@ -94,6 +95,7 @@ set(gui_HEADERS
         windows/registers/registersdock.h
         dialogs/savechanged/savechangeddialog.h
         dialogs/webeval/webevalconfigdialog.h
+        windows/webeval/webevaldock.h
         windows/editor/srceditor.h
         statictable.h
         windows/terminal/terminaldock.h
